requirements.txt
numpy
pyJWT
flask_login
requests